Complete Phase 1: Critical Infrastructure Hardening
Implement comprehensive type safety and CI/CD stabilization based on
PR #49 investigation findings stored in ChromaDB.

## Type Safety & CI/CD Stabilization
- Add automated pre-push hooks with type checking enforcement
- Create VS Code configuration for real-time type checking
- Implement type safety code review checklist
- Add CI/CD pipeline monitoring with automated alerts
- Enhance development workflow standardization

## Key Components Added
- .github/PULL_REQUEST_TEMPLATE/type_safety_checklist.md: Comprehensive PR review checklist
- .github/workflows/type-safety-monitoring.yml: Automated type safety monitoring
- docs/DEVELOPMENT_WORKFLOW.md: Complete development workflow guide
- scripts/setup_dev_environment.py: One-command development setup (renamed)
- scripts/configure_branch_protection.py: Branch protection automation (renamed)

## Enhanced Components
- CLAUDE.md: Updated with VS Code IDE configuration instructions
- scripts/verify.py: Added CI/CD modes (--ci, --json, --pre-push)

## Anti-Patterns Prevention
Prevents specific issues identified in PR #49 ChromaDB investigation:
- Optional attribute access without None checks
- Untyped fixture parameters
- Pydantic decorator type issues
- Missing generator return types

Note: Minor type checking issues remain in new scripts that will be
addressed in follow-up commits to maintain stability.

+++ tests/utils/helpers.py
@@ -238,7 +238,7 @@ class MockHelper:
         host: str = "localhost",
         port: int = 8000,
         ssl: bool = False,
-        timeout: float = 5.0,
+        timeout: int = 5,
         auth_token: str | None = None,
     ) -> ChromaDBConfig:
         """Create a mock ChromaDB configuration.
